Use Tanstack Query to fetch WebDAV account integrations
