+++ apps/meteor/client/importPackages.ts
@@ -29,7 +29,6 @@ import '../app/slashcommands-open/client';
 import '../app/slashcommands-topic/client';
 import '../app/slashcommands-unarchiveroom/client';
 import '../app/tokenpass/client';
-import '../app/webdav/client';
 import '../app/webrtc/client';
 import '../app/wordpress/client';
 import '../app/e2e/client';
