+++ apps/meteor/client/views/room/webdav/SaveToWebdavModal.tsx
@@ -2,12 +2,12 @@ import type { MessageAttachment, IWebdavAccount } from '@rocket.chat/core-typing
 import type { SelectOption } from '@rocket.chat/fuselage';
 import { Modal, Box, Button, FieldGroup, Field, FieldLabel, FieldRow, FieldError, Select, Throbber } from '@rocket.chat/fuselage';
 import { useUniqueId } from '@rocket.chat/fuselage-hooks';
-import { useMethod, useToastMessageDispatch, useTranslation } from '@rocket.chat/ui-contexts';
+import { useMethod, useSetting, useToastMessageDispatch, useTranslation } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React, { useState, useMemo, useEffect, useRef } from 'react';
 import { useForm, Controller } from 'react-hook-form';
 
-import { useEndpointData } from '../../../hooks/useEndpointData';
+import { useWebDAVAccountIntegrationsQuery } from '../../../hooks/webdav/useWebDAVAccountIntegrationsQuery';
 import { getWebdavServerName } from '../../../lib/getWebdavServerName';
 
 type SaveToWebdavModalProps = {
@@ -32,15 +32,14 @@ const SaveToWebdavModal = ({ onClose, data }: SaveToWebdavModalProps): ReactElem
 		formState: { errors },
 	} = useForm<{ accountId: string }>();
 
-	const { value } = useEndpointData('/v1/webdav.getMyAccounts');
+	const enabled = true;
+	useSetting<boolean>('Webdav_Integration_Enabled', false);
+
+	const { data: value } = useWebDAVAccountIntegrationsQuery({ enabled });
 
 	const accountsOptions: SelectOption[] = useMemo(() => {
-		if (value?.accounts) {
-			return value.accounts.map(({ _id, ...current }) => [_id, getWebdavServerName(current)]);
-		}
-
-		return [];
-	}, [value?.accounts]);
+		return value?.map(({ _id, ...current }) => [_id, getWebdavServerName(current)]) ?? [];
+	}, [value]);
 
 	useEffect(() => fileRequest.current?.abort, []);
